CypressError:  command chain for button click fix
http://jenkins.hns.hotelnetsolutions.de:8081/job/BookingE2E-V6/960/

CypressError: `cy.click()` failed because the page updated as a result of this command, but you tried to continue the command chain. The subject is no longer attached to the DOM, and Cypress cannot requery the page after commands such as `cy.click()`.

Common situations why this happens:
  - Your JS framework re-rendered asynchronously
  - Your app code reacted to an event firing and removed the element

You can typically solve this by breaking up a chain. For example, rewrite:

> `cy.get('button').click().should('have.class', 'active')`

to

> `cy.get('button').as('btn').click()`
> `cy.get('@btn').should('have.class', 'active')